Add raw dump option
Added option -r, --raw_dump. The default is no raw dump.
When the option is on no pcap global header is added.

This allows simpler processing of output frames
for programs that are interested mainly on the RTU
frame that is inside pcap frame. No need to skip
the global header.

In raw_dump mode also no human readable output is done
to stderr.